No more `Slackify.load_handlers`
This was introduced as we were doing an auth test to get the bot id.
We'll no longer do this moving forward. We'll start using a
list of whitelisted bot id. Any bot id not in the list will be
ignored. Since the bot id of the app will not be in this list, we
don't have to deal with auth'ing and storing the bot id.